pyproject.toml
[project]
name = "claude-agent-sdk-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "claude-agent-sdk>=0.1.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
